requirements.txt
Django
djangorestframework
decorator
django-environ
django-filter
psycopg2-binary